solve hex errors

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: damateos <damateos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/24 18:36:39 by damateos          #+#    #+#             */
-/*   Updated: 2024/05/13 22:38:23 by damateos         ###   ########.fr       */
+/*   Updated: 2024/05/15 23:27:17 by damateos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,12 @@
 
 # include <stdarg.h>
 # include <unistd.h>
+# include <stdint.h>
 # include "libft/libft.h"
 
 int	ft_printf(char const *format_str, ...);
 int	pf_putunbr(unsigned int n);
-int	pf_puthex(unsigned int n, int is_upper);
+int	pf_puthex(uintptr_t n, int is_upper);
 int	pf_putptr(void *ptr);
 int	pf_putchar(char c);
 int	pf_putstr(char *s);
